customer account rework
I've added a bunch of new files to rework how a customer is created and logs in. First, i'm slowly cutting out the customer classes that the professor gave us out. I've replaced them with my own version of customer, for now named TempCustomer, and added login functionality for it. This commit will allow a customer to register, have their data stored, and then log in. the next commit will be individual login pages for the customer, merchant, driver and staff.